Add in liboctomap-dev key.
The situation with octomap and ROS is complicated.

As it stands, we effectively have a "vendor" package of
octomap released into Humble, Iron, Jazzy, and Rolling.
However, this is a problem on Jazzy and Rolling on Ubuntu 24.04
since the ABI between what we have released, and what is
released in the operating system as "liboctomap-dev" is
different.  That means that using it with downstream
packages just doesn't work.

The intent here is to add this key in, and then remove
octomap from all of Humble, Iron, Jazzy, and Rolling.
This will avoid the ABI problem, and just use the system
package everywhere.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>